feat(북마크추가) : Board찾기 수행 북마크에 추가할 Board를 찾기 BookmarkController.java===== 요청에 들어온 board가 있는지 확인
BoardService.java=====
board를 찾는 메서드 없으면 NOTExist를 리턴한다.

BookmarkRequest.java====
int형식을 long으로 변경

+++ src/main/java/com/modureview/dto/request/BookmarkRequest.java
@@ -1,5 +1,5 @@
 package com.modureview.dto.request;
 
-public record BookmarkRequest(int userId, int boardId) {
+public record BookmarkRequest(Long userId, Long boardId) {
 
 }
